requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.4
pydantic==2.7.2
pydantic-core==2.18.3
python-decouple==3.8
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.12.0
